copaw-skill-platform/api/requirements.txt
fastapi==0.115.12
uvicorn==0.34.0
sqlalchemy==2.0.38
psycopg2-binary==2.9.10
python-jose[cryptography]==3.4.0
passlib[bcrypt]==1.7.4
pydantic==2.10.6
email-validator==2.2.0
